v0.25.4.16

+++ Core.cpk/Scripts/Systems/LandClaim/LandClaimSystemPurgeMechanic.cs
@@ -2,10 +2,12 @@ namespace AtomicTorch.CBND.CoreMod.Systems.LandClaim
 {
     using System.Collections.Generic;
     using System.Diagnostics;
+    using AtomicTorch.CBND.CoreMod.StaticObjects.Structures;
     using AtomicTorch.CBND.CoreMod.StaticObjects.Structures.Barrels;
     using AtomicTorch.CBND.CoreMod.StaticObjects.Structures.Crates;
     using AtomicTorch.CBND.CoreMod.StaticObjects.Structures.Manufacturers;
     using AtomicTorch.CBND.CoreMod.StaticObjects.Structures.TradingStations;
+    using AtomicTorch.CBND.CoreMod.Systems.PvE;
     using AtomicTorch.CBND.GameApi.Data.Items;
     using AtomicTorch.CBND.GameApi.Data.World;
     using AtomicTorch.CBND.GameApi.Scripting;
@@ -18,17 +20,38 @@ namespace AtomicTorch.CBND.CoreMod.Systems.LandClaim
     [UsedImplicitly]
     public static class LandClaimSystemPurgeMechanic
     {
+        private static void ServerDeleteAllStructures(
+            HashSet<IStaticWorldObject> staticWorldObjects,
+            out int objectsDeletedCount)
+        {
+            var world = Api.Server.World;
+            objectsDeletedCount = 0;
+
+            foreach (var worldObject in staticWorldObjects)
+            {
+                var protoStaticWorldObject = worldObject.ProtoStaticWorldObject;
+                if (!(protoStaticWorldObject is IProtoObjectStructure)
+                    || LandClaimSystem.SharedIsObjectInsideAnyArea(worldObject))
+                {
+                    continue;
+                }
+
+                world.DestroyObject(worldObject);
+                objectsDeletedCount++;
+            }
+        }
+
         // Please note:
         // usually this method finishes quickly so it's not spread across multiple frames.
-        private static void ServerPurgeBuildingsInDestroyedLandClaimArea(
+        private static void ServerPurgeStructuresInDestroyedLandClaimArea(
             IStaticWorldObject landClaimStructure,
             in RectangleInt areaBounds)
         {
+            var logger = Api.Logger;
             var stopwatch = Stopwatch.StartNew();
             var tempObjects = new HashSet<IStaticWorldObject>();
             var serverWorld = Api.Server.World;
             var serverItems = Api.Server.Items;
-            var logger = Api.Logger;
 
             for (var x = areaBounds.X; x < areaBounds.X + areaBounds.Width; x++)
             for (var y = areaBounds.Y; y < areaBounds.Y + areaBounds.Height; y++)
@@ -48,26 +71,35 @@ namespace AtomicTorch.CBND.CoreMod.Systems.LandClaim
                 }
             }
 
+            var objectsDeletedCount = 0;
             var purgedContainersCount = 0;
+
             foreach (var worldObject in tempObjects)
             {
-                switch (worldObject.ProtoStaticWorldObject)
+                var protoStaticWorldObject = worldObject.ProtoStaticWorldObject;
+                if (!(protoStaticWorldObject is IProtoObjectStructure)
+                    || LandClaimSystem.SharedIsObjectInsideAnyArea(worldObject))
                 {
-                    case IProtoObjectCrate _ when SharedIsObjectInFreeArea():
+                    continue;
+                }
+
+                switch (protoStaticWorldObject)
+                {
+                    case IProtoObjectCrate _:
                     {
                         var privateState = worldObject.GetPrivateState<ObjectCratePrivateState>();
                         PurgeContainer(privateState.ItemsContainer);
                         break;
                     }
 
-                    case IProtoObjectTradingStation _ when SharedIsObjectInFreeArea():
+                    case IProtoObjectTradingStation _:
                     {
                         var privateState = worldObject.GetPrivateState<ObjectTradingStationPrivateState>();
                         PurgeContainer(privateState.StockItemsContainer);
                         break;
                     }
 
-                    case IProtoObjectBarrel _ when SharedIsObjectInFreeArea():
+                    case IProtoObjectBarrel _:
                     {
                         var privateState = worldObject.GetPrivateState<ProtoBarrelPrivateState>();
                         privateState.LiquidAmount = 0;
@@ -76,21 +108,24 @@ namespace AtomicTorch.CBND.CoreMod.Systems.LandClaim
                         break;
                     }
 
-                    case IProtoObjectManufacturer _ when SharedIsObjectInFreeArea():
+                    case IProtoObjectManufacturer _:
                     {
                         var privateState = worldObject.GetPrivateState<ObjectManufacturerPrivateState>();
                         PurgeManufacturerContainers(privateState);
                         break;
                     }
                 }
+            }
 
-                bool SharedIsObjectInFreeArea()
-                    => !LandClaimSystem.SharedIsObjectInsideAnyArea(worldObject);
+            if (PveSystem.ServerIsPvE)
+            {
+                // on PvE server, delete all the structures within the decayed land claim area
+                ServerDeleteAllStructures(tempObjects, out objectsDeletedCount);
             }
 
             stopwatch.Stop();
             logger.Important(
-                $"Land claim destroyed: {landClaimStructure}. Items containers purged: {purgedContainersCount}. Time spent: {stopwatch.Elapsed.TotalMilliseconds}ms");
+                $"Land claim destroyed: {landClaimStructure}. Objects deleted: {objectsDeletedCount}. Items containers purged: {purgedContainersCount}. Time spent: {stopwatch.Elapsed.TotalMilliseconds}ms");
 
             void PurgeContainer(IItemsContainer container)
             {
@@ -128,9 +163,19 @@ namespace AtomicTorch.CBND.CoreMod.Systems.LandClaim
 
             private static void ServerObjectLandClaimDestroyedHandler(
                 IStaticWorldObject landClaimStructure,
-                RectangleInt areaBounds)
+                RectangleInt areaBounds,
+                bool isDestroyedByPlayers)
             {
-                ServerPurgeBuildingsInDestroyedLandClaimArea(landClaimStructure, areaBounds);
+                if (!PveSystem.ServerIsPvE
+                    && isDestroyedByPlayers)
+                {
+                    Api.Logger.Important(
+                        $"Land claim destroyed after timer due to players' attack: {landClaimStructure}. No items were purged.");
+                    return;
+                }
+
+                ServerPurgeStructuresInDestroyedLandClaimArea(landClaimStructure,
+                                                              areaBounds);
             }
         }
     }
